requirements.txt
alembic==1.8.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
boto3==1.26.57
botocore==1.29.57
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
fastapi-pagination==0.10.0
greenlet==1.1.3
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
jmespath==1.0.1
Mako==1.2.2
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
zipp==3.8.1
